Add Java implementation.
Better than the C++ version, probably due to the Java
version using a mutable board. C++ as it stands uses produces
a copy each time a move is made. Seems to be faster to make/unmake
moves, than to make a copy. Also better for memory, if that ever
was a concern. Might want to clean up the C++ version.